Agregados casos de testeo desde Postman, Aagregadas funcionlidades de matricuación y careras, Agregados DTO.

+++ estudiantesMySQL/src/main/java/com/estudiantes/controller/StudentControllerJpa.java
@@ -63,7 +63,7 @@ public class StudentControllerJpa {
     //2d) recuperar un estudiante, en base a su número de libreta universitaria.
     //TESTEADO EN POSTMAN
     @GetMapping("/bookNumber/{bookNumber}")
-    public Student getStudentByBookNumber(@PathVariable Long bookNumber) {
+    public Student getStudentByBookNumber(@PathVariable Integer bookNumber) {
     	return studentService.getStudentByBookNumber(bookNumber);
     }
